fix: payment controller bugs & new fields

+++ internal/domain/entities/expense_category/expense_category_repository.go
@@ -5,11 +5,11 @@ import (
 )
 
 type Repository interface {
-	Count(ctx context.Context) (int64, error)
+	Count(ctx context.Context) (uint, error)
 	GetAll(ctx context.Context) ([]*ExpenseCategory, error)
 	GetPaginated(ctx context.Context, limit, offset int, sortBy []string) ([]*ExpenseCategory, error)
-	GetByID(ctx context.Context, id int64) (*ExpenseCategory, error)
+	GetByID(ctx context.Context, id uint) (*ExpenseCategory, error)
 	Create(ctx context.Context, user *ExpenseCategory) error
 	Update(ctx context.Context, user *ExpenseCategory) error
-	Delete(ctx context.Context, id int64) error
+	Delete(ctx context.Context, id uint) error
 }
